Using pagination in get_dataset_metadata_export()
Using pagination in get_dataset_metadata_export(), by adding the limit and offset parameters to the API call to get the metadata of all versions of the dataset. The limit and offset params were added in Dataverse software v6.1.

I'm forced to do this because when I use the the "List Versions of a Dataset" endpoint for datasets on Harvard Dataverse and I don't specify a limit, a limit of 10 versions is applied.

Not all installations seem to do this, e.g. ADA. See the dataset at https://doi.org/10.4225/87/PJO7GB